ci: add some notes, npm `new_release` script & let Travis create draft releases
the Github release is not supposed to have any files other than source code
which is already part of the tag. Unfortunately, the Travis CI release
uploads a bunch of files, and it's difficult to tell Travis not to
upload any files, so we'll just create draft releases and manually
delete the uploaded files on Github.